Added floor and background sprites
Added floor and background sprites. Layout is for testing/example
purposes and is not final. Background prop sprites have 2D polygon
colliders so the characters can collide with them after being pushed out
of the ring. I wasn't sure which changes to ignore, so I just pushed all
of them.